Workshop-3/core-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
pytest
python-dotenv
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
